refactor: id 타입 일치화

+++ src/main/java/com/plog/domain/post/repository/PostRepository.java
@@ -25,5 +25,5 @@ import org.springframework.stereotype.Repository;
  */
 
 @Repository
-public interface PostRepository extends JpaRepository<Post, Integer> {
+public interface PostRepository extends JpaRepository<Post, Long> {
 }
